Input and Database needed
Still need to figure out how to bind the input string (from the input box) to an object of the type Company. This allows us to add a company (name, rating) to the array of companies.

Still no clue on how to work with a database, so the list can be stored permanently.